// src/data/enums/permissions.ts
export enum Permissions {
  SUPER_ADMIN = 'SUPER_ADMIN',
  AGENCY = 'AGENCY',
  AGENCY_ADMIN = 'AGENCY_ADMIN',
  TALENT_MANAGER = 'TALENT_MANAGER',
  PAGE = 'PAGE',
  USER_ADMIN = 'USER_ADMIN',
}
export enum PermissionAction {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
  READ = 'READ',
  MANAGE = 'MANAGE',
}